Move templates and routes into magazine folder
