+++ packages/global/components/magazine/templates/index.marko
