+++ packages/global/components/magazine/templates/publication.marko
@@ -1,4 +1,4 @@
-import issueFragment from "../../graphql/fragments/magazine-issue-archive";
+import issueFragment from "../graphql/fragments/magazine-issue-archive";
 
 $ const { pagination: p, req } = out.global;
 $ const { id, pageNode } = data;
