+++ packages/global/routes/index.js
@@ -2,7 +2,7 @@ const htmlSitemap = require('@parameter1/base-cms-marko-web-html-sitemap/routes'
 const dynamicPage = require('./dynamic-page');
 const feed = require('./feed');
 // const identityX = require('./identity-x');
-const magazine = require('./magazine');
+const magazine = require('../components/magazine/routes');
 const nativeX = require('./native-x');
 const printContent = require('./print-content');
 const publicFiles = require('./public-files');
